ConsoleMsg

+++ ProtocolNew.cpp
@@ -10983,17 +10983,17 @@ ConsoleMsg::ConsoleMsg() : ServerPacket(ServerPacketID_ConsoleMsg /* 24 */), Cha
 ConsoleMsg::ConsoleMsg(clsByteQueue* buffer) : ServerPacket(ServerPacketID_ConsoleMsg /* 24 */) {
     buffer->ReadByte(); /* PacketID */
     Chat = buffer->ReadUnicodeString();
-    FontIndex = buffer->ReadInteger();
+    FontIndex = buffer->ReadByte();
 
 }
 
-ConsoleMsg::ConsoleMsg(const std::string& Chat_, std::int16_t FontIndex_) : ServerPacket(ServerPacketID_ConsoleMsg /* 24 */), Chat(Chat_), FontIndex(FontIndex_) {
+ConsoleMsg::ConsoleMsg(const std::string& Chat_, std::uint8_t FontIndex_) : ServerPacket(ServerPacketID_ConsoleMsg /* 24 */), Chat(Chat_), FontIndex(FontIndex_) {
 }
 
 void ConsoleMsg::serialize(clsByteQueue* buffer) const {
     buffer->WriteByte(ServerPacketID_ConsoleMsg); /* PacketID: 24 */
     buffer->WriteUnicodeString(Chat);
-    buffer->WriteInteger(FontIndex);
+    buffer->WriteByte(FontIndex);
 
 }
